add pkgs for better overview
