--- pkg/mocks/mocks.go
@@ -1,1 +1,0 @@
-package mocks
